genai/genai-app-demo/01-prompt-engineering-demo/requirements.txt
aws-cdk-lib==2.134.0
constructs>=10.0.0,<11.0.0
cdk-nag==2.28.60
cdk-ecr-deployment==3.0.40
boto3==1.34.59
langchain==0.1.12
streamlit==1.32.2
requests==2.31.0